COM projection seems to work
