+++ freemocap_blender_addon/blender_ui/operators/_add_com_vertical_projection.py
@@ -1,7 +1,6 @@
-
-from freemocap_blender_addon.blender_ui.ui_utilities import add_com_vertical_projection
-
 import bpy
+
+from freemocap_blender_addon.core_functions.com_bos.add_com_vertical_projection import add_com_vertical_projection
 
 class FREEMOCAP_OT_add_com_vertical_projection(bpy.types.Operator):
     bl_idname = 'freemocap._add_com_vertical_projection'
@@ -10,11 +9,11 @@ class FREEMOCAP_OT_add_com_vertical_projection(bpy.types.Operator):
     bl_options = {'REGISTER', 'UNDO_GROUPED'}
 
     def execute(self, context):
-
         print("Adding COM Vertical Projection.......")
-
+        data_parent_object = context.scene.freemocap_properties.data_parent_empty
         # Add COM Vertical Projection
-        add_com_vertical_projection(neutral_color=context.scene.freemocap_ui_properties.com_vertical_projection_neutral_color,
+        add_com_vertical_projection(data_parent_object=data_parent_object,
+                                    neutral_color=context.scene.freemocap_ui_properties.com_vertical_projection_neutral_color,
                                     in_bos_color=context.scene.freemocap_ui_properties.com_vertical_projection_in_bos_color,
                                     out_bos_color=context.scene.freemocap_ui_properties.com_vertical_projection_out_bos_color)
 
@@ -22,3 +21,5 @@ class FREEMOCAP_OT_add_com_vertical_projection(bpy.types.Operator):
         context.scene.freemocap_ui_properties.show_com_vertical_projection = True
 
         return {'FINISHED'}
+
+
