+++ freemocap_blender_addon/blender_ui/ui_utilities.py
@@ -21,19 +21,6 @@ except ImportError:
     print("Replace the blender installation folder to match your installation.")
 
 
-# Function to hide (or unhide) Blender objects
-def hide_objects(data_object: bpy.types.Object,
-                 hide: bool = True,
-                 hide_children_not_parent: bool = False, ) -> None:
-    if hide_children_not_parent:
-        for child_object in data_object.children:
-            # Hide child object
-            child_object.hide_set(hide)
-            # Execute the function recursively
-            hide_objects(child_object, hide, hide_children_not_parent)
-    else:
-        data_object.hide_set(hide)
-
 
 # Function to draw a vector for debbuging purposes
 def draw_vector(origin, angle, name):
@@ -93,194 +80,8 @@ def graham_scan(points):
 
 # Function to toggle the visibility of the output elements.
 # It uses the parent_pattern to operate on the correct elements
-def toggle_element_visibility(self,
-                              context,
-                              panel_property: str,
-                              parent_pattern: str,
-                              toggle_children_not_parent: bool,)->None:
-
-    for data_object in bpy.data.objects:
-        if re.search(parent_pattern, data_object.name):
-            hide_objects(data_object,
-                         not bool(self[panel_property]),
-                         toggle_children_not_parent)
-
-# Function to toggle the visibility of the motion paths
-def toggle_motion_path(self,
-                       context,
-                       panel_property: str,
-                       data_object: str,
-                       show_line: bool = True,
-                       line_thickness: int = 6,
-                       use_custom_color: bool = False,
-                       line_color: tuple = (0.5, 1.0, 0.8),
-                       frames_before: int = 10,
-                       frames_after: int = 10,
-                       frame_step: int = 1,
-                       show_frame_numbers: bool = False,
-                       show_keyframes: bool = False,
-                       show_keyframe_number: bool = False)->None:
-
-    # Deselect all objects
-    bpy.ops.object.select_all(action='DESELECT')
-
-    # Get reference to the object
-    obj = bpy.data.objects[data_object]
-
-    # Select the object
-    obj.select_set(True)
-
-    # Set the object as active object
-    bpy.context.view_layer.objects.active = obj
-
-    if bool(self[panel_property]):
-        # Calculate paths
-        bpy.ops.object.paths_calculate(display_type='CURRENT_FRAME', range='SCENE')
-        # Set motion path properties for the specific object
-        if obj.motion_path:
-            obj.motion_path.lines = show_line
-            obj.motion_path.line_thickness = line_thickness
-            obj.motion_path.use_custom_color = use_custom_color
-            obj.motion_path.color = line_color
-            obj.animation_visualization.motion_path.frame_before = frames_before
-            obj.animation_visualization.motion_path.frame_after = frames_after
-            obj.animation_visualization.motion_path.frame_step = frame_step
-            obj.animation_visualization.motion_path.show_frame_numbers = show_frame_numbers
-            obj.animation_visualization.motion_path.show_keyframe_highlight = show_keyframes
-            obj.animation_visualization.motion_path.show_keyframe_numbers = show_keyframe_number
-    else:
-        bpy.ops.object.paths_clear(only_selected=True)
 
 # Function to add the COM Vertical Projection as COM mesh copy locked to the z axis floor plane
-def add_com_vertical_projection(neutral_color: tuple,
-                                in_bos_color: tuple,
-                                out_bos_color: tuple)->None:
-
-    # Deselect all objects
-    bpy.ops.object.select_all(action='DESELECT')
-
-    # Add a sphere mesh to the scene
-    bpy.ops.mesh.primitive_uv_sphere_add(enter_editmode=False,
-                                         align='WORLD',
-                                         location=(0, 0, 0),
-                                         scale=(0.05, 0.05, 0.05))
-    
-    # Change the name of the sphere mesh
-    bpy.context.active_object.name = "COM_Vertical_Projection"
-
-    # Get the mesh object
-    COM_mesh = bpy.data.objects["COM_Vertical_Projection"]
-
-    # Parent the sphere mesh to the capture origin empty
-    for data_object in bpy.data.objects:
-        if re.search(r'_origin\Z', data_object.name):
-            # Set the object as the COM vertical projection parent
-            COM_mesh.parent = data_object
-            break
-
-    # Create the mesh materials
-    bpy.ops.material.new()
-    bpy.data.materials["Material"].name = "COM_Vertical_Projection_Neutral"
-    COM_mesh.data.materials.append(bpy.data.materials["COM_Vertical_Projection_Neutral"])
-    # Change the color of the material
-    bpy.data.materials["COM_Vertical_Projection_Neutral"].node_tree.nodes["Principled BSDF"].inputs[0].default_value = neutral_color
-    bpy.data.materials["COM_Vertical_Projection_Neutral"].diffuse_color = neutral_color
-
-    bpy.ops.material.new()
-    bpy.data.materials["Material"].name = "COM_Vertical_Projection_In_BOS"
-    COM_mesh.data.materials.append(bpy.data.materials["COM_Vertical_Projection_In_BOS"])
-    # Change the color of the material
-    bpy.data.materials["COM_Vertical_Projection_In_BOS"].node_tree.nodes["Principled BSDF"].inputs[0].default_value = in_bos_color
-    bpy.data.materials["COM_Vertical_Projection_In_BOS"].diffuse_color = in_bos_color
-
-    bpy.ops.material.new()
-    bpy.data.materials["Material"].name = "COM_Vertical_Projection_Out_BOS"
-    COM_mesh.data.materials.append(bpy.data.materials["COM_Vertical_Projection_Out_BOS"])
-    # Change the color of the material
-    bpy.data.materials["COM_Vertical_Projection_Out_BOS"].node_tree.nodes["Principled BSDF"].inputs[0].default_value = out_bos_color
-    bpy.data.materials["COM_Vertical_Projection_Out_BOS"].diffuse_color = out_bos_color
-
-    # Create a Geometry Nodes modifier to switch the material depending on the BOS intersection
-    bpy.ops.node.new_geometry_nodes_modifier()
-
-    # Change the name of the geometry node
-    COM_mesh.modifiers[0].name = "Geometry Nodes_" + COM_mesh.name
-   
-    # Get the node tree and change its name
-    node_tree = bpy.data.node_groups[0]
-    node_tree.name = "Geometry Nodes_" + COM_mesh.name
-
-    # Get the Input and Output nodes
-    input_node = node_tree.nodes["Group Input"]
-    output_node = node_tree.nodes["Group Output"]
-
-    # Add the Material node for the Neutral Material
-    material_neutral_node = node_tree.nodes.new(type="GeometryNodeInputMaterial")
-    # Change the node name
-    material_neutral_node.name = "Material Neutral"
-    # Assign the material to the node
-    node_tree.nodes["Material Neutral"].material = bpy.data.materials["COM_Vertical_Projection_Neutral"]
-
-    # Add the Material node for the In BOS Material
-    material_in_bos_node = node_tree.nodes.new(type="GeometryNodeInputMaterial")
-    # Change the node name
-    material_in_bos_node.name = "Material In BOS"
-    # Assign the material to the node
-    node_tree.nodes["Material In BOS"].material = bpy.data.materials["COM_Vertical_Projection_In_BOS"]
-
-    # Add the Material node for the Out BOS Material
-    material_out_bos_node = node_tree.nodes.new(type="GeometryNodeInputMaterial")
-    # Change the node name
-    material_out_bos_node.name = "Material Out BOS"
-    # Assign the material to the node
-    node_tree.nodes["Material Out BOS"].material = bpy.data.materials["COM_Vertical_Projection_Out_BOS"]
-
-    # Add a Switch Node for the In-Out BOS materials
-    in_out_bos_switch_node = node_tree.nodes.new(type='GeometryNodeSwitch')
-    # Change the node name
-    in_out_bos_switch_node.name = "In-Out BOS Switch"
-
-    # Change the input type of the in_out_bos switch node
-    node_tree.nodes["In-Out BOS Switch"].input_type = 'MATERIAL'
-
-    # Add a Switch Node for the BOS visible status
-    bos_visible_switch_node = node_tree.nodes.new(type='GeometryNodeSwitch')
-    # Change the node name
-    bos_visible_switch_node.name = "BOS Visible Switch"
-
-    # Change the input type of the in_out_bos switch node
-    node_tree.nodes["BOS Visible Switch"].input_type = 'MATERIAL'
-
-    # Add a Set Material Node
-    set_material_node =  node_tree.nodes.new(type="GeometryNodeSetMaterial")
-
-    # Connect the material nodes to the two inputs of the switch node
-    node_tree.links.new(material_out_bos_node.outputs["Material"], in_out_bos_switch_node.inputs["False"])
-    node_tree.links.new(material_in_bos_node.outputs["Material"], in_out_bos_switch_node.inputs["True"])
-
-    # Connect the neutral material node to the bos visible switch node
-    node_tree.links.new(material_neutral_node.outputs["Material"], bos_visible_switch_node.inputs["False"])
-
-    # Connect the in_out_bos switch node to the bos visible switch node
-    node_tree.links.new(in_out_bos_switch_node.outputs["Output"], bos_visible_switch_node.inputs["True"])
-
-    # Connect the bos visible switch node to the set material node
-    node_tree.links.new(bos_visible_switch_node.outputs["Output"], set_material_node.inputs["Material"])
-
-    # Connect the input node to the set material node
-    node_tree.links.new(input_node.outputs["Geometry"], set_material_node.inputs["Geometry"])
-
-    # Connect the set material node to the output node
-    node_tree.links.new(set_material_node.outputs["Geometry"], output_node.inputs["Geometry"])
-
-    # Add a copy location constraint to the COM vertical projection
-    bpy.ops.object.constraint_add(type='COPY_LOCATION')
-
-    # Set the copy location target as the Center of Mass mesh
-    COM_mesh.constraints["Copy Location"].target = bpy.data.objects["center_of_mass_mesh"]
-
-    # Disable the constraint on the z axis constraint
-    COM_mesh.constraints["Copy Location"].use_z = False
 
 # Function to add angle meshes, set a copy location constraint to a joint
 def add_angle_meshes(points: dict,
@@ -806,20 +607,3 @@ def add_base_of_support(z_threshold: float,
     return
 
 
-def update_motion_path(self, context, data_object: str):
-    toggle_motion_path(
-        self,
-        context,
-        panel_property='motion_path_' + data_object,
-        data_object=data_object,
-        show_line=self.motion_path_show_line,
-        line_thickness=self.motion_path_line_thickness,
-        use_custom_color=self.motion_path_use_custom_color,
-        line_color=self.motion_path_line_color,
-        frames_before=self.motion_path_frames_before,
-        frames_after=self.motion_path_frames_after,
-        frame_step=self.motion_path_frame_step,
-        show_frame_numbers=self.motion_path_show_frame_numbers,
-        show_keyframes=self.motion_path_show_keyframes,
-        show_keyframe_number=self.motion_path_show_keyframe_number
-    )
